Rename per-chart color strategy modules
JIRA: RAIL-1815
