+++ libs/sdk-ui-charts/src/highcharts/chartTypes/bubbleChart/bubbleChartColoring.ts
