+++ libs/sdk-ui-charts/src/highcharts/chartTypes/bulletChart/bulletChartColoring.ts
