+++ libs/sdk-ui-charts/src/highcharts/chartTypes/heatmap/heatmapColoring.ts
