+++ libs/sdk-ui-charts/src/highcharts/chartTypes/scatterPlot/scatterPlotColoring.ts
