+++ libs/sdk-ui-charts/src/highcharts/chartTypes/treemap/treemapColoring.ts
